requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
psycopg2-binary
python-jose
python-multipart
fastapi_jwt_auth
fastapi-mail
pytz
